requirements.txt
openai
streamlit
arxiv
python-dotenv
pandas
